+++ src/main/java/com/feelory/feelory_backend/global/file/FileConfig.java
@@ -1,4 +1,4 @@
-package com.feelory.feelory_backend.global.config;
+package com.feelory.feelory_backend.global.file;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
@@ -8,7 +8,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import java.nio.file.Paths;
 
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+public class FileConfig implements WebMvcConfigurer {
 
     @Value("${file.upload.path}")
     private String uploadPath;
